Take version-rolling mask from stratum

+++ components/asic/include/asic.h
@@ -12,6 +12,6 @@ void ASIC_set_job_difficulty_mask(GlobalState * GLOBAL_STATE, uint8_t mask);
 void ASIC_send_work(GlobalState * GLOBAL_STATE, void * next_job);
 void ASIC_set_version_mask(GlobalState * GLOBAL_STATE, uint32_t mask);
 bool ASIC_set_frequency(GlobalState * GLOBAL_STATE, float target_frequency);
-double ASIC_get_asic_job_frequency_ms(GlobalState * GLOBAL_STATE);
+float ASIC_get_asic_job_frequency_ms(GlobalState * GLOBAL_STATE);
 
 #endif // ASIC_H
